220710 17:04
BOARD
 : userInfo API 수정
 : 유효성 검사
 : 관리자 => 배송기사 공지사항 접근가능
 : 글 작성 버튼
   - 기존 발생가능한 에러 : state가 업데이트 되기전에 권한이 없어도 작성이 가능함
   - 수정

MAIN
 : aside 수정
 : header animation 수정

DELIVERY
 : logout

+++ src/main/java/com/ez/launer/user/model/DriverAllVO.java
@@ -18,7 +18,7 @@ public class DriverAllVO {
 	private Timestamp outDate;
 	private int accInfoNo;
 	private String bank;
-	private int accNum;
+	private long accNum;
 	private String accHolder;
 	private String randomPwd;
 }
